Change FontSize return type from u16 to u32

+++ src/the_dude_to_human/database/dude_types.h
@@ -202,7 +202,7 @@ struct LongArrayField {
         return fmt::format("[{}]", array);
     }
 
-    u16 FontSize() const {
+    u32 FontSize() const {
         if (data.size() < 2) {
             return 0;
         }
